+++ examples/expo-example/.storybook/main.ts
@@ -17,10 +17,10 @@ const main: StorybookConfig = {
     // '../components/**/*.storiesof.?(ts|tsx|js|jsx)',
   ],
   addons: [
-    '@storybook/addon-ondevice-notes',
     '@storybook/addon-ondevice-controls',
     '@storybook/addon-ondevice-backgrounds',
     '@storybook/addon-ondevice-actions',
+    '@storybook/addon-ondevice-notes',
   ],
 };
