+++ examples/expo-example/.storybook/storybook.requires.ts
@@ -1,16 +1,11 @@
 /* do not change this file, it is auto generated by storybook. */
 
-import {
-  start,
-  prepareStories,
-  getProjectAnnotations,
-  updateView,
-} from "@storybook/react-native";
+import { start, updateView } from "@storybook/react-native";
 
-import "@storybook/addon-ondevice-notes/register";
 import "@storybook/addon-ondevice-controls/register";
 import "@storybook/addon-ondevice-backgrounds/register";
 import "@storybook/addon-ondevice-actions/register";
+import "@storybook/addon-ondevice-notes/register";
 
 const normalizedStories = [
   {
